chore: remove native webpack config
- Native widgets now use native webpack config from PIW tools directly
