--- packages/pluggableWidgets/feedback-native/webpack.config.dev.js
@@ -1,1 +1,0 @@
-module.exports = require("../../../configs/webpack.dev.native.js");
