--- packages/pluggableWidgets/listview-swipe-native/webpack.config.prod.js
@@ -1,1 +1,0 @@
-module.exports = require("../../../configs/webpack.prod.native.js");
